+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - New command `/q debug ingame`
   - Debug logging
     - now with 10 minutes history
+- NotifyIO "totemIO" 
 ### Changed
 - changed package names from `pl.betoncraft.betonquest` to `org.betonquest.betonquest`
 ### Deprecated